pyproject.toml
[project]
name = "book-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpraw>=7.8.1",
    "bs4>=0.0.2",
    "openai>=1.58.1",
    "pynytimes>=0.10.0",
    "python-dotenv>=1.0.1",
    "python-telegram-bot>=21.9",
]
